bigquery: first cut at fast query path.
This PR is conservative, and only tries to implement fast
query path for consumers who construct a Query object and then invoke
Read() on it.  I have no sense for how common this usage is, vs users
who invoke Run() on the Query object to get a Job, and call Read() on
that.

It is missing some things like parameter support, and the fact that the
tests don't fail indicates we need more test coverage for this kind of
thing.